# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weed Logger APIs",
        "description": "Weed Logger APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/forget/password/request": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Forget Password Email",
                "description": "Enter your email here to get link to update password.",
                "operationId": "Forget Password Email",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Please check your email for password reset."
                    },
                    "422": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/update-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update Password",
                "description": "Enter your email and password here to update password.",
                "operationId": ".Update Password",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "You have successfully changed your password"
                    },
                    "422": {
                        "description": "Invalid token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/installations": {
            "get": {
                "tags": [
                    "Installation"
                ],
                "summary": "List of all Installations",
                "description": "Get list of all Installations here",
                "operationId": "List of all Installations",
                "responses": {
                    "200": {
                        "description": "All Installations"
                    },
                    "403": {
                        "description": "You are not authorized."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Installation"
                ],
                "summary": "Add New Installation",
                "description": "Add new installation here",
                "operationId": "Add installations",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "companyId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "companyId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Installation Register Successfully !!"
                    },
                    "403": {
                        "description": "You are not authorized."
                    },
                    "422": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/installations/{id}": {
            "get": {
                "tags": [
                    "Installation"
                ],
                "summary": "Show Installation By Id",
                "description": "Show installation by id",
                "operationId": "Show Installations By Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Show Installations by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Installations  show successfully with this id."
                    },
                    "403": {
                        "description": "You are not authorized."
                    },
                    "422": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Installation"
                ],
                "summary": "Update Installation By Id",
                "description": "Update installation here using id",
                "operationId": "Update Installation By Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update installation by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "companyId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "companyId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Installation Updated Successfully"
                    },
                    "403": {
                        "description": "You are not authorized."
                    },
                    "422": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}